+++ packages/integration/lib/src/views/edit_short_video/edit_short_video_screen.dart
@@ -377,9 +377,10 @@ class _LMFeedEditShortVideoScreenState
       },
       style: LMFeedButtonStyle(
         borderRadius: 4,
+        gap: 4,
         padding: EdgeInsets.symmetric(
           horizontal: 8,
-          vertical: 6,
+          vertical: 4,
         ),
         margin: EdgeInsets.only(
           right: 8,
@@ -423,7 +424,10 @@ class _LMFeedEditShortVideoScreenState
         ValueListenableBuilder(
             valueListenable: _postValidationNotifier,
             builder: (context, value, child) {
-              return _defPostButton();
+              return _screenBuilder.editButtonBuilder(
+                context,
+                _defPostButton(),
+              );
             }),
       ],
       style: LMFeedAppBarStyle(