+++ packages/integration/lib/src/views/select_topic/configurations/style.dart
@@ -1,0 +1,3 @@
+class LMFeedTopicSelectScreenStyle {
+  const LMFeedTopicSelectScreenStyle();
+}